requirements.txt
numpy
torchvision
zarr
matplotlib
tqdm
astropy
sunpy
imageio
opencv-python
